--- usvm-ts/src/test/kotlin/org/usvm/util/ModifiersExt.kt
@@ -1,8 +1,0 @@
-package org.usvm.util
-
-import org.jacodb.ets.model.EtsModifier
-import org.jacodb.ets.model.EtsModifiers
-
-fun EtsModifiers.toStrings(): List<String> {
-    return EtsModifier.entries.filter { hasModifier(it) }.map { it.string }
-}
